[test] NFC: Extract out `NullEditorConsumer`
